feat: add docs suspense (#22)
