+++ src/App.tsx
@@ -4,20 +4,22 @@ import {
   RouterProvider,
   Route,
 } from 'react-router-dom';
-import { FC, Suspense } from 'react';
+import { FC } from 'react';
 import { ThemeProvider } from '@mui/material/styles';
-import { SyncLoader } from 'react-spinners';
 import { Home, Auth, EditorPage, NotFound } from '@/pages';
 import { Layout } from '@/components/layout';
 import { Login } from '@/components/login';
 import { SignUp } from '@/components/sign-up';
 import { PrivateRoute } from '@/components/private-route';
-import { LocalStorageKeys, RoutePath } from './utils/enum';
+import {
+  LocalStorageKeys,
+  RoutePath,
+  customDark,
+  customLight,
+  changeCssRootVariables,
+  getLocalStorage,
+} from './utils';
 import { useAppSelector } from './hooks/redux';
-import { override } from './utils/const';
-import { customDark, customLight } from './utils/customTheme';
-import { changeCssRootVariables } from './utils/change-css-root';
-import { getLocalStorage } from './utils/local-storage-utils';
 
 const router = createBrowserRouter(
   createRoutesFromElements(
@@ -41,9 +43,7 @@ export const App: FC = () => {
   changeCssRootVariables(theme);
   return (
     <ThemeProvider theme={isNightMode ? customDark : customLight}>
-      <Suspense fallback={<SyncLoader cssOverride={override} color="#768fa3" size={25} />}>
-        <RouterProvider router={router} />
-      </Suspense>
+      <RouterProvider router={router} />
     </ThemeProvider>
   );
 };
