+++ src/components/alert/index.ts
@@ -1,0 +1,1 @@
+export { Alert } from './alert';
