+++ src/components/editor/editor-response-section/editor-response-section.scss
