+++ src/components/editor/editor-section/editor-section.scss
