+++ src/components/editor/index.ts
@@ -1,1 +1,3 @@
-export { Editor } from './editor';
+export { EditorDocsSection } from './editor-docs-section';
+export { EditorResponseSection } from './editor-response-section';
+export { EditorSection } from './editor-section';
