+++ src/locales/ru.json
@@ -30,8 +30,6 @@
   "editor-page.docs-short": "Дока",
   "email.required": "Поле не может быть пустым",
   "email.invalid": "Электронная почта недействительна",
-  "email.length": "Электронная почта должна быть длиннее 8 символов",
-  "email.digit": "Электронная почта должна содержать хотя бы одну цифру",
   "pswd.length": "Пароль должен быть длиннее 8 символов",
   "pswd.required": "Поле не может быть пустым",
   "pswd.digit": "Пароль должен содержать хотя бы одну цифру",
@@ -42,6 +40,7 @@
   "exist.email": "Этот электронный адрес уже занят",
   "some.error": "Что-то пошло не так",
   "editor-page.error-schema": "Ошибка загрузки схемы!",
+  "editor-page.docs-loading": "Загрузка документации",
   "error-boundary.title": "Ошибка",
   "error-boundary.text": "Для перезагрузки страницы нажмите кнопку.",
   "error-boundary.button": "Перезагрузить"
