+++ src/slices/index.ts
@@ -1,0 +1,1 @@
+export { changeCssRootVariables } from './change-css-root';
